refactor(data-collection): Decouple collection logic from data management
This commit refactors the box office data collection process to better
adhere to the Single Responsibility Principle. The logic for batch
processing has been moved from the low-level collector to the high-level
data management class.

Problem/Reason:
- The `BoxOfficeCollector` was previously responsible for both fetching
  data for a single movie and managing the entire batch collection
  process for a dataset.
- This created tight coupling, as the low-level collector had to
  understand the structure of `MovieData`, `BoxOfficeProgressFile`, and
  how to persist data, making it difficult to reuse and test.

Adjustments:
- `BoxOfficeCollector` is now a pure, low-level tool. Its sole public
  responsibility is `fetch_single_movie_data`, which returns data in
  memory without any side effects like file I/O or progress tracking.
- The `download_box_office_data_for_movies` method and the
  `BoxOfficeProgressFile` helper class have been removed from
  `box_office_collector.py`.
- `Dataset.collect_box_office` now acts as the high-level orchestrator.
  It manages progress tracking, iterates through movies, calls the
  `BoxOfficeCollector` as a tool, and handles data persistence through
  its repository.

Impact/Contrast:
- Before: `BoxOfficeCollector` was a monolithic class handling both web
  scraping and application-specific workflow logic.
- After: `BoxOfficeCollector` is a reusable, independent web scraping
  utility. `Dataset` is the application-specific workflow manager. This
  separation improves modularity, testability, and maintainability.

BREAKING CHANGE: The public interface of `BoxOfficeCollector` has changed.
The `download_box_office_data_for_movies` method has been removed.
To perform batch collection for a dataset, users must now instantiate a
`Dataset` object and call its `collect_box_office` method. For single,
in-memory queries, `BoxOfficeCollector.fetch_single_movie_data` remains
the correct entry point.
