+++ src/data_collection/box_office_collector.py
@@ -2,7 +2,7 @@ import json
 import tempfile
 from logging import Logger
 from pathlib import Path
-from typing import Callable, Final, Literal, Optional, TypeAlias, TypedDict, Tuple
+from typing import Final, Literal, Optional, Tuple, TypeAlias
 from urllib.parse import quote
 
 from selenium.common.exceptions import (
@@ -18,218 +18,28 @@ from selenium.webdriver.support.expected_conditions import (
     text_to_be_present_in_element,
     visibility_of_element_located
 )
-from tqdm import tqdm
 from urllib3.exceptions import ReadTimeoutError
-from yaml import YAMLError
 
-from src.core.constants import Constants
 from src.core.logging_manager import LoggingManager
 from src.data_collection.browser import Browser, ElementLocator
 from src.data_handling.box_office import BoxOffice
-from src.data_handling.file_io import CsvFile, YamlFile
-from src.data_handling.movie_collections import MovieData
 
 DownloadFinishCondition: TypeAlias = Browser.DownloadFinishCondition
 PageChangeCondition: TypeAlias = Browser.PageChangeCondition
 WaitingCondition: TypeAlias = Browser.WaitingCondition
 
 
-class BoxOfficeProgressEntry(TypedDict):
-    """
-    Represents the structure of a single entry in the box office download progress file.
-
-    :ivar id: The unique identifier for the movie.
-    :ivar url: The URL of the movie's box office data page.
-    :ivar file_path: The local path where the movie's box office data is saved.
-    """
-    id: int
-    url: str
-    file_path: str
-
-
-class BoxOfficeProgressFile(CsvFile):
-    """
-    Handles read/write operations for the box office download progress CSV file.
-
-    This class extends CsvFile to manage a progress file that tracks the download
-    status (URL, saved file path) for each movie.
-
-    :ivar HEADER: A tuple defining the CSV header fields: ('id', 'url', 'file_path').
-    """
-
-    HEADER: Final[tuple[str, str, str]] = ('id', 'url', 'file_path')
-
-    def __init__(self, path: Path, encoding: str = Constants.DEFAULT_ENCODING):
-        """
-        Initializes the BoxOfficeProgressFile handler.
-
-        :param path: The path to the progress CSV file.
-        :param encoding: The encoding of the file, defaults to 'utf-8'.
-        """
-        super().__init__(path=path, encoding=encoding, header=self.HEADER)
-        self.__logger: Logger = LoggingManager().get_logger('root')
-
-    def save(self, data: list[BoxOfficeProgressEntry]) -> None:
-        """
-        Saves a list of progress entries to the CSV file.
-
-        This method writes the provided data, ensuring the parent directory exists.
-        It uses the class's predefined HEADER for the CSV field names.
-
-        :param data: A list of ``BoxOfficeProgressEntry`` dictionaries to save.
-        """
-        super().save(data=data)
-        self.__logger.info(f"Successfully saved {len(data)} progress entries to '{self.path}'.")
-        return
-
-    def load(self, row_factory: Optional[Callable[[dict[str, str]], any]] = None) -> list[BoxOfficeProgressEntry]:
-        """
-        Loads and parses data from the progress CSV file.
-
-        This method overrides the parent ``CsvFile.load`` to use a specific internal
-        row factory (``_progress_entry_factory``) for converting rows into
-        ``BoxOfficeProgressEntry`` dictionaries. A warning is logged if an external
-        ``row_factory`` is provided, as it will be ignored.
-
-        :param row_factory: This parameter is ignored. A warning will be logged if it is provided.
-        :returns: A list of parsed ``BoxOfficeProgressEntry`` objects. Returns an
-                  empty list if the file does not exist or is empty.
-        :raises Exception: Propagates exceptions from the underlying CSV loading
-                           process, except for ``FileNotFoundError``.
-        """
-        if not self.path.exists():
-            self.__logger.info(f"Progress file not found at '{self.path}'. Returning empty list.")
-            return []
-
-        if row_factory is not None:
-            self.__logger.warning(
-                "BoxOfficeProgressFile.load was called with a 'row_factory' argument, "
-                "but it will use its internal '_progress_entry_factory' for conversion."
-            )
-
-        try:
-            loaded_entries: list[Optional[BoxOfficeProgressEntry]] = super().load(
-                row_factory=self._progress_entry_factory)
-            processed_data: list[BoxOfficeProgressEntry] = [entry for entry in loaded_entries if entry is not None]
-            return processed_data
-        except FileNotFoundError:
-            self.__logger.error(f"FileNotFoundError during load after exists() check for '{self.path}'.")
-            return []
-        except Exception as e:
-            self.__logger.error(f"Error loading progress file '{self.path}': {e}", exc_info=True)
-            raise
-
-    @staticmethod
-    def _progress_entry_factory(row: dict[str, str]) -> Optional[BoxOfficeProgressEntry]:
-        """
-        Converts a raw CSV row into a structured ``BoxOfficeProgressEntry``.
-
-        This factory function validates the input row, ensuring the 'id' field
-        exists and is a valid integer. If the row is invalid, it logs a warning
-        and returns ``None``.
-
-        :param row: A dictionary representing a single row from the CSV file.
-        :returns: A ``BoxOfficeProgressEntry`` instance if the row is valid,
-                  otherwise ``None``.
-        """
-        logger: Logger = LoggingManager().get_logger(
-            BoxOfficeProgressFile.__name__)
-        movie_id_str: Optional[str] = row.get('id')
-        url_str: str = row.get('url', '')
-        file_path_str: str = row.get('file_path', '')
-
-        if movie_id_str is None:
-            logger.warning(f"Skipping progress entry due to missing 'id': {row}")
-            return None
-
-        try:
-            movie_id: int = int(movie_id_str)
-        except ValueError:
-            logger.warning(
-                f"Skipping progress entry due to invalid 'id' format: '{movie_id_str}' in {row}")
-            return None
-
-        return BoxOfficeProgressEntry(id=movie_id, url=url_str, file_path=file_path_str)
-
-    def initialize_from_movies(self, movies: list[MovieData]) -> None:
-        """
-        Creates and initializes the progress file from a list of movies.
-
-        This method generates an initial progress entry for each movie, setting the
-        'id' from the movie data and leaving 'url' and 'file_path' empty.
-        It will overwrite the progress file if it already exists.
-
-        :param movies: A list of ``MovieData`` objects to use for initialization.
-        """
-        initial_data: list[BoxOfficeProgressEntry] = [
-            BoxOfficeProgressEntry(id=movie.id, url='', file_path='') for movie in movies
-        ]
-        self.save(data=initial_data)
-        self.__logger.info(f"Initialized progress file '{self.path}' with {len(initial_data)} entries.")
-        return
-
-    def update_entry(self, movie_id: int, update_field: Literal['url', 'file_path'], new_value: str) -> None:
-        """
-        Updates a single field for a specific movie entry in the progress file.
-
-        This method reads the entire progress file, finds the entry matching the
-        ``movie_id``, modifies the specified ``update_field`` with the ``new_value``,
-        and then writes the entire dataset back to the file.
-
-        :param movie_id: The ID of the movie entry to update.
-        :param update_field: The name of the field to update (either 'url' or 'file_path').
-        :param new_value: The new value to set for the field.
-        :raises ValueError: If the ``movie_id`` is not found in the progress file or
-                            if ``update_field`` is not a valid field name.
-        :raises FileNotFoundError: If the progress file does not exist when an update
-                                   is attempted.
-        """
-        current_progress: list[BoxOfficeProgressEntry] = self.load()
-        if not current_progress and not self.exists:
-            raise FileNotFoundError(
-                f"Progress file '{self.path}' not found. Cannot update entry for movie ID {movie_id}.")
-
-        target_entry: Optional[BoxOfficeProgressEntry] = None
-        entry_index: int = -1
-        for i, entry in enumerate(current_progress):
-            if entry.get('id') == movie_id:
-                target_entry = entry
-                entry_index = i
-                break
-
-        if target_entry is None:
-            msg: str = f"Movie ID {movie_id} not found in progress file '{self.path}'. Cannot update."
-            self.__logger.error(msg)
-            raise ValueError(msg)
-
-        if update_field == 'url':
-            current_progress[entry_index]['url'] = new_value
-        elif update_field == 'file_path':
-            current_progress[entry_index]['file_path'] = new_value
-        else:
-            invalid_field_msg: str = f"Invalid update_field: '{update_field}'. Must be 'url' or 'file_path'."
-            self.__logger.error(invalid_field_msg)
-            raise ValueError(invalid_field_msg)
-
-        self.save(data=current_progress)
-        self.__logger.debug(f"Updated {update_field} for movie ID {movie_id} in progress file.")
-        return
-
-
 class BoxOfficeCollector:
     """
-    Collects box office data from a specific website.
-
+    Collects box office data for a single movie from a specific website.
 
     This class encapsulates the logic for navigating the website, searching for
-    movies, and downloading their box office data. It is designed to be used
+    a movie, and downloading its box office data. It is designed to be used
     as a context manager to ensure the underlying browser instance is properly
     managed.
 
-    It supports two primary use cases:
-    1. Fetching data for a single movie and returning it in memory.
-    2. Batch downloading data for multiple movies, saving the results to disk
-       and tracking progress to allow for resumption.
+    The primary method, `fetch_single_movie_data`, handles the entire
+    process for one movie and returns the data in memory.
 
     :ivar __download_mode: The mode for downloading data (e.g., weekly or weekend).
     :ivar __logger: Logger instance for logging messages.
@@ -567,12 +377,9 @@ class BoxOfficeCollector:
             raise
         return
 
-    def __search_and_fetch_box_office(self,
-                                      movie_name: str,
-                                      movie_id: Optional[int] = None,
-                                      known_url: Optional[str] = None,
-                                      trying_times: int = 3) -> \
-        Tuple[Optional[list[BoxOffice]], Optional[str]]:
+    def __search_and_fetch_box_office(
+        self, movie_name: str, movie_id: Optional[int] = None, known_url: Optional[str] = None, trying_times: int = 3
+    ) -> Tuple[Optional[list[BoxOffice]], Optional[str]]:
         """
         Handles the complete workflow for fetching a single movie's box office data.
 
@@ -637,112 +444,33 @@ class BoxOfficeCollector:
                             return None, None
         return None, None
 
-    def download_box_office_data_for_movie(self, movie_name: str) -> list[BoxOffice]:
+    def fetch_single_movie_data(
+        self, movie_name: str, movie_id: Optional[int] = None, known_url: Optional[str] = None
+    ) -> Tuple[Optional[list[BoxOffice]], Optional[str]]:
         """
         Fetches box office data for a single movie and returns it in memory.
 
-        This is a simplified public method for retrieving data for one movie without
-        persisting it to disk or using a progress file. It must be called within
-        the ``with`` context of the collector.
+        This is the primary public method for retrieving data for one movie.
+        It must be called within the ``with`` context of the collector.
 
         :param movie_name: The name of the movie to fetch data for.
-        :returns: A list of ``BoxOffice`` objects representing the movie's data.
-        :raises RuntimeError: If the data cannot be fetched after multiple attempts.
+        :param movie_id: The unique ID of the movie, used for logging purposes.
+        :param known_url: An optional, pre-existing URL for the movie's page to
+                          attempt direct navigation, potentially speeding up the process.
+        :returns: A tuple containing:
+                  - A list of ``BoxOffice`` objects representing the movie's data, or ``None`` if fetching fails.
+                  - The final URL of the movie's page, or ``None`` if navigation fails.
         """
         self._check_browser_active()
 
-        box_office_data, _ = self.__search_and_fetch_box_office(
-            movie_name=movie_name, movie_id=None, known_url=None)
+        # The core logic is already in __search_and_fetch_box_office. We just call it.
+        box_office_data, movie_url = self.__search_and_fetch_box_office(
+            movie_name=movie_name, movie_id=movie_id, known_url=known_url
+        )
+
         if box_office_data is None:
-            error_message: str = f"Failed to download box office data for movie '{movie_name}' after multiple attempts."
-            self.__logger.error(error_message)
-            raise RuntimeError(error_message)
-        return box_office_data
+            self.__logger.error(
+                f"Failed to download box office data for movie '{movie_name}' (ID: {movie_id}) after multiple attempts."
+            )
 
-    def download_box_office_data_for_movies(self,
-                                            multiple_movie_data: list[MovieData],
-                                            data_folder: Path) -> None:
-        """
-        Downloads and saves box office data for a list of movies.
-
-        This method iterates through a list of movies, downloading and saving the
-        box office data for each one to the specified ``data_folder``. It uses a
-        progress file (``download_progress.csv``) within that folder to track
-        completed downloads, allowing the process to be resumed. If a download
-        fails for a movie, an empty data file is created to prevent re-attempts.
-
-        :param multiple_movie_data: A list of ``MovieData`` objects to process.
-        :param data_folder: The target directory to save the YAML data files and
-                            the progress CSV file.
-        """
-        self._check_browser_active()
-        self.__logger.info(f"Starting batch download to '{data_folder}'.")
-        data_folder.mkdir(parents=True, exist_ok=True)
-
-        progress_file: BoxOfficeProgressFile = BoxOfficeProgressFile(
-            path=data_folder / "download_progress.csv")
-
-        if not progress_file.exists:
-            self.__logger.info(f"Progress file '{progress_file.path}' not found. Initializing.")
-            progress_file.initialize_from_movies(movies=multiple_movie_data)
-
-        self.__logger.info(f"Loading progress file '{progress_file.path}' into memory...")
-        all_progress_entries: list[BoxOfficeProgressEntry] = progress_file.load()
-        progress_map: dict[int, BoxOfficeProgressEntry] = {entry['id']: entry for entry in all_progress_entries}
-        self.__logger.info(f"Loaded {len(progress_map)} entries into progress map.")
-        try:
-            with tqdm(
-                total=len(multiple_movie_data), bar_format=Constants.STATUS_BAR_FORMAT, desc="Collecting Box Office"
-            ) as pbar:
-                for movie in multiple_movie_data:
-                    pbar.set_postfix_str(f"Movie: {movie.name[:30]}...", refresh=True)
-
-                    # Check progress to see if we can skip based on file_path
-                    progress: Optional[BoxOfficeProgressEntry] = progress_map.get(movie.id)
-                    if progress and progress.get('file_path') and Path(progress['file_path']).exists():
-                        self.__logger.info(f"Data for movie ID {movie.id} already exists. Skipping.")
-                        pbar.update(1)
-                        continue
-
-                    known_url: Optional[str] = progress.get('url') if progress else None
-
-                    # Fetch data, passing the progress_file for potential URL optimization
-                    box_office_data, movie_url = self.__search_and_fetch_box_office(
-                        movie_name=movie.name, movie_id=movie.id, known_url=known_url)
-
-                    # Persist data and update progress if fetch was successful
-                    if box_office_data and movie_url:
-                        movie.update_box_office(data=box_office_data, update_method='REPLACE')
-                        saved_path: Path = movie.save_box_office(target_directory=data_folder)
-                        # Update progress file
-                        if progress:
-                            progress['url'] = movie_url
-                            progress['file_path'] = str(saved_path)
-                        self.__logger.info(f"Box office data for movie ID {movie.id} processed and saved.")
-                    else:
-                        # Handle failure by creating an empty file to prevent re-attempts
-                        empty_file_path: Path = data_folder / f"{movie.id}.yaml"
-                        self.__logger.warning(
-                            f"No box office data found for movie ID {movie.id}. Creating empty file."
-                        )
-                        try:
-                            YamlFile(path=empty_file_path).save([])
-                            # Update in-memory progress map
-                            if progress:
-                                progress['url'] = movie_url if movie_url else ''
-                                progress['file_path'] = str(empty_file_path)
-                        except (OSError, YAMLError) as e:
-                            self.__logger.error(
-                                f"Error creating empty file for movie ID {movie.id}: {e}", exc_info=True)
-
-                    pbar.update(1)
-        finally:
-            self.__logger.info("Collection loop finished. Saving all progress updates to disk...")
-            # Convert map back to list, preserving order if necessary (though order isn't critical here)
-            updated_progress_list: list[BoxOfficeProgressEntry] = list(progress_map.values())
-            try:
-                progress_file.save(data=updated_progress_list)
-            except Exception as e:
-                self.__logger.critical(
-                    f"Failed to save final progress to '{progress_file.path}': {e}", exc_info=True)
-        return
+        return box_office_data, movie_url
