+++ src/data_handling/dataset.py
@@ -2,11 +2,13 @@ from dataclasses import dataclass, field, replace, InitVar
 from logging import Logger
 from pathlib import Path
 from time import sleep
-from typing import Final, Literal, Optional
+from typing import Any, Callable, cast, Final, Literal, Optional, TypedDict
 
 from tqdm import tqdm
+from yaml import YAMLError
 
 from data_handling.database_client import DatabaseClient
+from src.core.constants import Constants
 from src.core.logging_manager import LoggingManager
 from src.core.project_config import DatabaseConfig, ProjectConfig, ProjectDatasetType, ProjectPaths
 from src.data_collection.box_office_collector import BoxOfficeCollector
@@ -18,6 +20,190 @@ from src.data_handling.repositories.repository import MovieRepository
 from src.data_handling.repositories.yaml_repository import YamlMovieRepository
 from src.data_handling.reviews import PublicReview
 from src.sentiment_analysis.llm_client import DailyRateLimitExceededError, LLMClient, LLMProvider
+
+
+class BoxOfficeProgressEntry(TypedDict):
+    """
+    Represents the structure of a single entry in the box office download progress file.
+
+    :ivar id: The unique identifier for the movie.
+    :ivar url: The URL of the movie's box office data page.
+    :ivar processed: The status of collection of the movie.
+    """
+    id: int
+    url: str
+    processed: bool
+
+
+class BoxOfficeProgressFile(CsvFile):
+    """
+    Handles read/write operations for the box office download progress CSV file.
+
+    This class extends CsvFile to manage a progress file that tracks the download
+    status (URL, processed) for each movie.
+
+    :ivar HEADER: A tuple defining the CSV header fields: ('id', 'url', 'processed').
+    """
+
+    HEADER: Final[tuple[str, str, str]] = ('id', 'url', 'processed')
+
+    def __init__(self, path: Path, encoding: str = Constants.DEFAULT_ENCODING):
+        """
+        Initializes the BoxOfficeProgressFile handler.
+
+        :param path: The path to the progress CSV file.
+        :param encoding: The encoding of the file, defaults to 'utf-8'.
+        """
+        super().__init__(path=path, encoding=encoding, header=self.HEADER)
+        self.__logger: Logger = LoggingManager().get_logger('root')
+
+    def save(self, data: list[BoxOfficeProgressEntry]) -> None:
+        """
+        Saves a list of progress entries to the CSV file.
+
+        This method writes the provided data, ensuring the parent directory exists.
+        It uses the class's predefined HEADER for the CSV field names.
+
+        :param data: A list of ``BoxOfficeProgressEntry`` dictionaries to save.
+        """
+        super().save(data=cast(list[dict[Any, Any]], cast(object, data)))
+        self.__logger.info(f"Successfully saved {len(data)} progress entries to '{self.path}'.")
+        return
+
+    def load(self, row_factory: Optional[Callable[[dict[str, str]], any]] = None) -> list[BoxOfficeProgressEntry]:
+        """
+        Loads and parses data from the progress CSV file.
+
+        This method overrides the parent ``CsvFile.load`` to use a specific internal
+        row factory (``_progress_entry_factory``) for converting rows into
+        ``BoxOfficeProgressEntry`` dictionaries. A warning is logged if an external
+        ``row_factory`` is provided, as it will be ignored.
+
+        :param row_factory: This parameter is ignored. A warning will be logged if it is provided.
+        :returns: A list of parsed ``BoxOfficeProgressEntry`` objects. Returns an
+                  empty list if the file does not exist or is empty.
+        :raises Exception: Propagates exceptions from the underlying CSV loading
+                           process, except for ``FileNotFoundError``.
+        """
+        if not self.path.exists():
+            self.__logger.info(f"Progress file not found at '{self.path}'. Returning empty list.")
+            return []
+
+        if row_factory is not None:
+            self.__logger.warning(
+                "BoxOfficeProgressFile.load was called with a 'row_factory' argument, "
+                "but it will use its internal '_progress_entry_factory' for conversion."
+            )
+
+        try:
+            loaded_entries: list[Optional[BoxOfficeProgressEntry]] = super().load(
+                row_factory=self._progress_entry_factory)
+            processed_data: list[BoxOfficeProgressEntry] = [entry for entry in loaded_entries if entry is not None]
+            return processed_data
+        except FileNotFoundError:
+            self.__logger.error(f"FileNotFoundError during load after exists() check for '{self.path}'.")
+            return []
+        except Exception as e:
+            self.__logger.error(f"Error loading progress file '{self.path}': {e}", exc_info=True)
+            raise
+
+    @staticmethod
+    def _progress_entry_factory(row: dict[str, str]) -> Optional[BoxOfficeProgressEntry]:
+        """
+        Converts a raw CSV row into a structured ``BoxOfficeProgressEntry``.
+
+        This factory function validates the input row, ensuring the 'id' field
+        exists and is a valid integer. If the row is invalid, it logs a warning
+        and returns ``None``.
+
+        :param row: A dictionary representing a single row from the CSV file.
+        :returns: A ``BoxOfficeProgressEntry`` instance if the row is valid,
+                  otherwise ``None``.
+        """
+        logger: Logger = LoggingManager().get_logger(
+            BoxOfficeProgressFile.__name__)
+        movie_id_str: Optional[str] = row.get('id')
+        url_str: str = row.get('url', '')
+        process_str: str = row.get('processed', '')
+
+        if movie_id_str is None:
+            logger.warning(f"Skipping progress entry due to missing 'id': {row}")
+            return None
+
+        try:
+            movie_id: int = int(movie_id_str)
+        except ValueError:
+            logger.warning(
+                f"Skipping progress entry due to invalid 'id' format: '{movie_id_str}' in {row}")
+            return None
+
+        processed_value: bool = process_str.strip().lower() == 'true'
+
+        return BoxOfficeProgressEntry(id=movie_id, url=url_str, processed=processed_value)
+
+    def initialize_from_movies(self, movies: list[MovieData]) -> None:
+        """
+        Creates and initializes the progress file from a list of movies.
+
+        This method generates an initial progress entry for each movie, setting the
+        'id' from the movie data and leaving 'url' and 'file_path' empty.
+        It will overwrite the progress file if it already exists.
+
+        :param movies: A list of ``MovieData`` objects to use for initialization.
+        """
+        initial_data: list[BoxOfficeProgressEntry] = [
+            BoxOfficeProgressEntry(id=movie.id, url='', processed=False) for movie in movies
+        ]
+        self.save(data=initial_data)
+        self.__logger.info(f"Initialized progress file '{self.path}' with {len(initial_data)} entries.")
+        return
+
+    def update_entry(self, movie_id: int, update_field: Literal['url', 'file_path'], new_value: str) -> None:
+        """
+        Updates a single field for a specific movie entry in the progress file.
+
+        This method reads the entire progress file, finds the entry matching the
+        ``movie_id``, modifies the specified ``update_field`` with the ``new_value``,
+        and then writes the entire dataset back to the file.
+
+        :param movie_id: The ID of the movie entry to update.
+        :param update_field: The name of the field to update (either 'url' or 'file_path').
+        :param new_value: The new value to set for the field.
+        :raises ValueError: If the ``movie_id`` is not found in the progress file or
+                            if ``update_field`` is not a valid field name.
+        :raises FileNotFoundError: If the progress file does not exist when an update
+                                   is attempted.
+        """
+        current_progress: list[BoxOfficeProgressEntry] = self.load()
+        if not current_progress and not self.exists:
+            raise FileNotFoundError(
+                f"Progress file '{self.path}' not found. Cannot update entry for movie ID {movie_id}.")
+
+        target_entry: Optional[BoxOfficeProgressEntry] = None
+        entry_index: int = -1
+        for i, entry in enumerate(current_progress):
+            if entry.get('id') == movie_id:
+                target_entry = entry
+                entry_index = i
+                break
+
+        if target_entry is None:
+            msg: str = f"Movie ID {movie_id} not found in progress file '{self.path}'. Cannot update."
+            self.__logger.error(msg)
+            raise ValueError(msg)
+
+        if update_field == 'url':
+            current_progress[entry_index]['url'] = new_value
+        elif update_field == 'processed':
+            current_progress[entry_index]['processed'] = bool(new_value)
+        else:
+            invalid_field_msg: str = f"Invalid update_field: '{update_field}'. Must be 'url' or 'processed'."
+            self.__logger.error(invalid_field_msg)
+            raise ValueError(invalid_field_msg)
+
+        self.save(data=current_progress)
+        self.__logger.debug(f"Updated {update_field} for movie ID {movie_id} in progress file.")
+        return
 
 
 @dataclass(kw_only=True)
@@ -157,23 +343,34 @@ class Dataset:
         """
         self.__logger.info(f"Initializing dataset '{self.name}' from source '{source_csv.path}'.")
 
-        # 1. Check if occupied
-        if self.repository.is_storage_occupied():
-            self.__logger.warning(f"Storage for dataset '{self.name}' is already occupied. Skipping initialization.")
-            return
-
-        # 2. Setup Storage Infrastructure
-        if self.mode == 'DATABASE' and root_config:
-            # Create Database and Grant Permissions (Admin Task)
+        if self.mode == 'DATABASE':
             admin_client = DatabaseClient(config=root_config)
             with admin_client.connection() as db:
-                db.execute_statement(f"CREATE DATABASE IF NOT EXISTS {self.name}")
-                db.execute_statement(f"GRANT ALL ON {self.name}.* TO '{self._database_config['user']}'@'%'")
+                # Check for database existence without connecting to it
+                result = db.execute_statement(f"SHOW DATABASES LIKE '{self.name}'")
+                db_exists = bool(result)
+
+            if db_exists:
+                if self.repository.is_storage_occupied():
+                    self.__logger.warning(
+                        f"Storage for dataset '{self.name}' is already occupied. Skipping initialization.")
+                    return
+            else:
+                # DB does not exist, create it and grant permissions
+                with admin_client.connection() as db:
+                    db.execute_statement(f"CREATE DATABASE IF NOT EXISTS {self.name}")
+                    db.execute_statement(f"GRANT ALL ON {self.name}.* TO '{self._database_config['user']}'@'%'")
+
+        elif self.mode == 'YAML_FILE':
+            if self.repository.is_storage_occupied():
+                self.__logger.warning(
+                    f"Storage for dataset '{self.name}' is already occupied. Skipping initialization.")
+                return
 
         # Delegate schema/folder creation to repository
         self.repository.setup_storage()
 
-        # 3. Load and Prepare Data
+        # Load and Prepare Data
         try:
             source_data: list[dict[str, str]] = source_csv.load()
             if not source_data:
@@ -186,7 +383,7 @@ class Dataset:
                 if movie_name:
                     movies.append(MovieData(id=index, name=movie_name))
 
-            # 4. Save Data
+            # Save Data
             if movies:
                 self.repository.save_movies(movies)
                 self.__logger.info(f"Successfully initialized dataset '{self.name}' with {len(movies)} movies.")
@@ -318,20 +515,78 @@ class Dataset:
 
         self.__logger.info(
             f"Collecting box office data for {len(movies_to_collect_for)} movies in dataset '{self.name}'.")
+
+        self.box_office_folder_path.mkdir(parents=True, exist_ok=True)
+
+        progress_file: BoxOfficeProgressFile = BoxOfficeProgressFile(
+            path=self.box_office_folder_path / "download_progress.csv")
+
+        if not progress_file.exists:
+            self.__logger.info(f"Progress file '{progress_file.path}' not found. Initializing.")
+            progress_file.initialize_from_movies(movies=movies_to_collect_for)
+
+        self.__logger.info(f"Loading progress file '{progress_file.path}' into memory...")
+        all_progress_entries: list[BoxOfficeProgressEntry] = progress_file.load()
+        progress_map: dict[int, BoxOfficeProgressEntry] = {entry['id']: entry for entry in all_progress_entries}
+        self.__logger.info(f"Loaded {len(progress_map)} entries into progress map.")
+
         try:
-
             with BoxOfficeCollector(download_mode='WEEK') as collector:
-                # Note: Collector still expects MovieData objects, but now they are simpler.
-                # The collector might need to know WHERE to save.
-                # Currently, collector takes 'data_folder'.
-                collector.download_box_office_data_for_movies(multiple_movie_data=movies_to_collect_for,
-                                                              data_folder=self.box_office_folder_path)
+                with tqdm(
+                    total=len(movies_to_collect_for), bar_format=Constants.STATUS_BAR_FORMAT, desc="Collecting Box Office"
+                ) as pbar:
+                    for movie in movies_to_collect_for:
+                        pbar.set_postfix_str(f"Movie: {movie.name[:30]}...", refresh=True)
 
-            self.__logger.info(f"Box office collection process finished for dataset '{self.name}'. "
-                               f"The `movie_data` cache remains invalidated; reload to see updates.")
-        except Exception as e:
-            self.__logger.error(f"An error occurred during box office collection for dataset '{self.name}': {e}",
-                                exc_info=True)
+                        # Check progress to see if we can skip based on file_path
+                        progress: Optional[BoxOfficeProgressEntry] = progress_map.get(movie.id)
+                        if progress and progress.get('processed') is True:
+                            self.__logger.info(f"Data for movie ID {movie.id} already exists. Skipping.")
+                            pbar.update(1)
+                            continue
+
+                        known_url: Optional[str] = progress.get('url') if progress else None
+
+                        # Fetch data using the refactored collector
+                        box_office_data, movie_url = collector.fetch_single_movie_data(
+                            movie_name=movie.name, movie_id=movie.id, known_url=known_url)
+
+                        # Persist data and update progress if fetch was successful
+                        if box_office_data and movie_url:
+                            movie.update_box_office(data=box_office_data, update_method='REPLACE')
+                            self.repository.save_box_office(movie_id=movie.id, data=box_office_data)
+                            # Update progress file
+                            if progress:
+                                progress['url'] = movie_url
+                                progress['processed'] = True
+                            self.__logger.info(f"Box office data for movie ID {movie.id} processed and saved.")
+                        else:
+                            # Handle failure by creating an empty file to prevent re-attempts
+                            self.__logger.warning(
+                                f"No box office data found for movie ID {movie.id}."
+                            )
+                            try:
+                                # Update in-memory progress map
+                                if progress:
+                                    progress['url'] = movie_url if movie_url else ''
+                                    progress['processed'] = False
+                            except (OSError, YAMLError) as e:
+                                self.__logger.error(
+                                    f"Error creating empty file for movie ID {movie.id}: {e}", exc_info=True)
+
+                        pbar.update(1)
+        finally:
+            self.__logger.info("Collection loop finished. Saving all progress updates to disk...")
+            # Convert map back to list, preserving order if necessary (though order isn't critical here)
+            updated_progress_list: list[BoxOfficeProgressEntry] = list(progress_map.values())
+            try:
+                progress_file.save(data=updated_progress_list)
+            except Exception as e:
+                self.__logger.critical(
+                    f"Failed to save final progress to '{progress_file.path}': {e}", exc_info=True)
+
+        self.__logger.info(f"Box office collection process finished for dataset '{self.name}'. "
+                           f"The `movie_data` cache remains invalidated; reload to see updates.")
         return
 
     def collect_public_review(self, target_website: Literal['PTT', 'DCARD']) -> None:
